rename: 이동, EditProfile을 auth -> dashboard로
